Added highly generic version of reaction generator keyed off an omniplex.  It can do modifications, exchange small-mols (such as nucleotides) and can accept additional reactants and products.  This reaction generator should suffice for many kinase, phosphatase, and nuclotide hydrolysis reactions.
I am thinking about getting rid of the reaction generators that specifically
model nucleotide binding with modifications.  There is very little left that
they can do that cannot be done with this new generic reaction generator.

A generic reaction generator for mol features, similar to what I did above
with omni features, would be good, and it would remind people that the mol
feature is available (and cheaper than omni.)  Maybe I should just make a
single module with all the generic reaction generators in it?

This version still does not build, fully, only enough of moleculizer to verify
that the new constructs are essentially working.  Did not bump version number
for this checkin.  


git-svn-id: https://moleculizer.svn.sourceforge.net/svnroot/moleculizer/moleculizer/trunk@8 e2b6e7e4-f0f2-0310-8dec-eb45598a66c7