Removing boost/foreach.hpp dependencies from the demonstration code.
At this point, the only code that depends on boost/foreach.hpp is the test/ code that is located in various directories.  I have chosen not to remove this, as basically it is primarily needed by the developers, and it's just too useful to remove.  The second location is the python layer code, which isn't really used, although I should go check that out very soon.




git-svn-id: https://moleculizer.svn.sourceforge.net/svnroot/moleculizer/libmoleculizer/trunk@638 e2b6e7e4-f0f2-0310-8dec-eb45598a66c7